Include CSS, SASS, and image loaders in the webpack config
Right now, if I need to use assets like images, CSS, or fonts, Webpack won't load them and include them in the bundle.  To make it possible to continue using Quik alongside these things, add the ability for the user to specify additional loaders to be included in the Webpack config.